Remove buffer and rolling
These functions did not work properly. bufferConnect and
rollingConnect are fine, but the approach of wrapping a single machine
in isolation -- independent of its connections -- was mistaken.

A new test suite guarded by a Cabal flag (the flag’s name is ‘splot‘)
has been added to help visualize parallel execution of machines. This
test is thanks to Ben Sinclair.

Fixes #5